Reorganize and add ik utils
