+++ colab_notebook_utils/__init__.py
@@ -1,3 +1,4 @@
-from .utils import (filter_analysis_dicts, filter_pop_df, get_animation,
-                    get_available_joint_angles, get_joint_plot,
-                    get_relative_frame, list_chunks, save_figs_to_html)
+from .colab_notebook_backend import (filter_analysis_dicts, filter_pop_df,
+                                     get_animation, get_available_joint_angles,
+                                     get_joint_plot, get_relative_frame,
+                                     list_chunks, save_figs_to_html)
