+++ colab_notebook_utils/datatypes.py
@@ -1,0 +1,11 @@
+from enum import Enum
+
+
+class Functions(str, Enum):
+    VISUALIZATIONS = "colab_notebook_backend"
+    INVERSE_KINEMATICS = "toolbox_ik"
+
+
+class InvocationTypes(str, Enum):
+    SYNC = "RequestResponse"
+    ASYNC = "Event"
